Add setup.py for pip installation
